Proyecto

+++ projects/project3/main.js
@@ -1,6 +1,5 @@
-showNotes();
+addBtn = document.getElementById("addBtn");
 
-let addBtn = document.getElementById("addBtn");
 addBtn.addEventListener("click", function (e) {
   let addTxt = document.getElementById("addTxt"),
     notes = localStorage.getItem("notes");
@@ -15,6 +14,13 @@ addBtn.addEventListener("click", function (e) {
   localStorage.setItem("notes", JSON.stringify(notesObj));
   addTxt.value = "";
   showNotes();
+  Swal.fire({
+    position: "top-end",
+    icon: "success",
+    title: "Your note has been saved",
+    showConfirmButton: false,
+    timer: 1500,
+  });
 });
 
 function showNotes() {
@@ -48,14 +54,47 @@ function showNotes() {
 }
 
 function deleteNote(index) {
-  let notes = localStorage.getItem("notes");
-  if (notes == null) {
-    notesObj = [];
-  } else {
-    notesObj = JSON.parse(notes);
-  }
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: "btn btn-success",
+      cancelButton: "btn btn-danger",
+    },
+    buttonsStyling: false,
+  });
 
-  notesObj.splice(index, 1);
-  localStorage.setItem("notes", JSON.stringify(notesObj));
-  showNotes();
+  swalWithBootstrapButtons
+    .fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "No, cancel!",
+      reverseButtons: true,
+    })
+    .then((result) => {
+      if (result.isConfirmed) {
+        let notes = localStorage.getItem("notes");
+        if (notes == null) {
+          notesObj = [];
+        } else {
+          notesObj = JSON.parse(notes);
+        }
+
+        notesObj.splice(index, 1);
+        localStorage.setItem("notes", JSON.stringify(notesObj));
+        showNotes();
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        swalWithBootstrapButtons.fire(
+          "Cancelled",
+          "Your imaginary file is safe :)",
+          "error"
+        );
+      }
+    });
 }
+
+showNotes();
